% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newZ.R
\name{newZ}
\alias{newZ}
\title{Helper function for creating new vector based on resampling the original data.}
\usage{
newZ(tM, n_bins)
}
\arguments{
\item{tM}{The outputs from \code{transM}; a \code{list} with elements \code{xn} and \code{MM}.}

\item{n_bins}{The number of bins to use; passed from \code{muti}.}
}
\value{
A vector of resampled values.
}
\description{
\code{newZ} creates new vector based on transition matrix.
}
