% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolize.R
\name{symbolize}
\alias{symbolize}
\title{Convert a numeric vector into a symbolic vector.}
\usage{
symbolize(xy)
}
\arguments{
\item{xy}{An [n x 2] \code{matrix} or \code{data.frame} containing the two vectors of interest.}
}
\value{
An [(n-2) x 2] \code{matrix} of integer symbols that indicate whether the i-th value,
based on the i-1 and i+1 values, is a "trough" (=1), "decrease" (=2), "same" (=3), "increase" (=4), or "peak" (=5).
}
\description{
\code{symbolize} converts a numeric vector into a symbolic vector.
}

