% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transM.R
\name{transM}
\alias{transM}
\title{Estimates the transition matrix for confidence intervals on mutual information.}
\usage{
transM(x, n_bins)
}
\arguments{
\item{x}{A \code{vector} of values.}

\item{n_bins}{The number of bins for the entropy calculation}
}
\value{
A list with the following components:
\describe{
\item{\code{xn}}{An [n x 2] matrix of the original and discretized vectors.}
\item{\code{MM}}{Transition probability matrix from bin-i to bin-j.}
}
}
\description{
\code{mutual_info} is a helper function for estimating the transition matrix
used in creating resampled vectors for the (1 - alpha)% confidence
interval on the mutual info.
}

