% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muti.R
\name{muti}
\alias{muti}
\title{Calculate mutual information between two vectors and plot results.}
\usage{
muti(x, y, n_bins = NULL, sym = TRUE, lags = seq(-4, 4), mc = 100,
  alpha = 0.05, normal = FALSE)
}
\arguments{
\item{x}{First vector of data}

\item{y}{Second vector of data}

\item{n_bins}{[optional] Specify the number of bins to use for calculating
the entropy. If \code{NULL} (default), the data are discretized based on
the "Rice Rule" where \code{n_bins = ceiling(2*length(x)^(1/3))}.}

\item{sym}{Boolean indicator of whether to use symbolic dynamics (default).
If \code{FALSE}, the data are discretized based on \code{n_bins}.}

\item{lags}{Vector of \code{integer} indicating what lags to use. Can be
positive and/or negative.}

\item{mc}{The number of Monte Carlo simulations for estimating the
confidence interval on the mutual information.}

\item{alpha}{The alpha value for the upper (1-alpha)% confidence bound
on the mutual information (lower bound is 0).}

\item{normal}{Boolean indicator of whether to normalize the mutual
information based on the entropies. See 'Details."}
}
\value{
A \code{data.frame} with columns for lag (\code{lag}), mutual
  information between x & y (\code{MI_xy}), and the upper confidence bound
  (\code{MI_ci}). Note that the lower bound for MI is 0. Plots of x & y,
  their symbolic values (if desired), and the mutual information at multiple
  lags.
}
\description{
\code{muti} calculates the mutual information between two vectors and plots
the results at multiple lags.
}
\examples{
set.seed(123)

## number of time steps
TT <- 30
## create vector 1
x <- rnorm(TT)

## CASE 1: no mutual info at any lag
y <- rnorm(TT)
muti(x,y)

## CASE 2: mutual info at lag 1
y <- x + rnorm(TT,0,0.5)
muti(x,y)

}

