% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutual_info.R
\name{mutual_info}
\alias{mutual_info}
\title{Calculate the mutual information between two vectors.}
\usage{
mutual_info(su, normal)
}
\arguments{
\item{su}{An n x 2 \code{matrix} containing two discrete vectors.}

\item{normal}{A logical indicator as to whether the mutual information
should be normalized to [0,1].}
}
\value{
A scalar of class `numeric` of the (possibly normalized) mutual
  information.
}
\description{
\code{mutual_info} calculates the mutual information between two vectors. It
  is a helper function for \code{muti}.
}
