% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMI.R
\name{plotMI}
\alias{plotMI}
\title{Plot original data plus the mutual information at multiple lags.}
\usage{
plotMI(x, y, MI, sym, n_bins)
}
\arguments{
\item{x}{First vector of data}

\item{y}{Second vector of data}

\item{MI}{A \code{list} with the mutual information as output by \code{muti}.}

\item{sym}{A boolean indcator as to whether symbolic dynamics (TRUE) were used.}

\item{n_bins}{The number of bins used for computing the MI.}
}
\value{
Plots of original vectors, (potential) symbolic values, and the mutual information at multiple lags.
}
\description{
\code{plotMI} plots original data plus the mutual information at multiple
lags.
}

