% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMI.R
\name{plotMI}
\alias{plotMI}
\title{Plot the data and mutual information.}
\usage{
plotMI(x, y, MI, sym, n_bins, normal)
}
\arguments{
\item{x}{First vector of data.}

\item{y}{Second vector of data.}

\item{MI}{A \code{list} with the mutual information as output by \code{muti}.}

\item{sym}{A logical indicator as to whether symbolic dynamics were used.}

\item{n_bins}{The number of bins used for computing the MI.}

\item{normal}{A logical indicator as to whether the MI was normalized to [0,1].}
}
\value{
A 3-panel plot of original vectors (top), their discretized values
  (middle), and the mutual information at multiple lags (bottom).
}
\description{
\code{plotMI} plots the original data, the symbolic or discretized data, and
  the mutual information at multiple lags. It is a helper function for
  \code{muti}.
}
